+++ orx-color/src/commonMain/kotlin/spaces/ColorHSLUVa.kt
@@ -2,6 +2,7 @@ package org.openrndr.extra.color.spaces
 
 import org.openrndr.color.*
 import org.openrndr.math.Vector4
+import org.openrndr.math.map
 import org.openrndr.math.mixAngle
 import kotlin.math.*
 
@@ -10,10 +11,10 @@ private val m = arrayOf(
         doubleArrayOf(-0.96924363628087, 1.87596750150772, 0.041555057407175),
         doubleArrayOf(0.055630079696993, -0.20397695888897, 1.056971514242878))
 
-private val kappa = 903.2962962
-private val epsilon = 0.0088564516
+private const val kappa = 903.2962962
+private const val epsilon = 0.0088564516
 
-private fun getBounds(L: Double): List<DoubleArray>? {
+private fun getBounds(L: Double): List<DoubleArray> {
     val result = ArrayList<DoubleArray>()
     val sub1 = (L + 16).pow(3.0) / 1560896
     val sub2 = if (sub1 > epsilon) sub1 else L / kappa
@@ -48,11 +49,11 @@ private class Length(val length: Double) {
     val greaterEqualZero: Boolean = length >= 0
 }
 
-private fun maxSafeChromaForL(L100: Double): Double {
+internal fun maxSafeChromaForL(L100: Double): Double {
     val bounds = getBounds(L100)
     var min = Double.MAX_VALUE
     for (i in 0..1) {
-        val m1 = bounds!![i][0]
+        val m1 = bounds[i][0]
         val b1 = bounds[i][1]
         val line = doubleArrayOf(m1, b1)
         val x = intersectLineLine(line, doubleArrayOf(-1 / m1, 0.0))
@@ -62,11 +63,11 @@ private fun maxSafeChromaForL(L100: Double): Double {
     return min
 }
 
-fun maxChromaForLH(L100: Double, H: Double): Double {
+private fun maxChromaForLH(L100: Double, H: Double): Double {
     val hrad = H / 360 * PI * 2
     val bounds = getBounds(L100)
     var min = Double.MAX_VALUE
-    for (bound in bounds!!) {
+    for (bound in bounds) {
         val length: Length = lengthOfRayUntilIntersect(hrad, bound)
         if (length.greaterEqualZero) {
             min = min(min, length.length)
@@ -144,53 +145,7 @@ fun mix(left: ColorHSLUVa, right: ColorHSLUVa, x: Double): ColorHSLUVa {
             (1.0 - sx) * left.alpha + sx * right.alpha)
 }
 
-data class ColorXSLUVa(val x: Double, val s: Double, val l: Double, override val alpha: Double = 1.0):
-        ColorModel<ColorXSLUVa>,
-        HueShiftableColor<ColorXSLUVa>,
-        SaturatableColor<ColorXSLUVa>,
-        ShadableColor<ColorXSLUVa>,
-        AlgebraicColor<ColorXSLUVa> {
-
-    @Deprecated("Legacy alpha parameter name", ReplaceWith("alpha"))
-    val a = alpha
-
-    override fun shiftHue(shiftInDegrees: Double) = copy(x = x + (shiftInDegrees))
-
-    override fun shade(factor: Double) = copy(l = l * factor)
-
-    override fun saturate(factor: Double) = copy(s = s * factor)
-
-    override fun toRGBa(): ColorRGBa {
-        return toHSLUVa().toRGBa()
-    }
-
-    fun toHSLUVa(): ColorHSLUVa = ColorHSLUVa(xToHue(x), s, l, alpha)
-
-    override fun opacify(factor: Double) = copy(alpha = alpha * factor)
-
-    override fun minus(right: ColorXSLUVa) = copy(x = x - right.x, s = s - right.s, l = l - right.l, alpha = alpha - right.alpha)
-
-    override fun plus(right: ColorXSLUVa) = copy(x = x + right.x, s = s + right.s, l = l + right.l, alpha = alpha + right.alpha)
-
-    override fun times(scale: Double) = copy(x = x * scale, s = s * scale, l = l * scale, alpha = alpha * scale)
-
-    override fun mix(other: ColorXSLUVa, factor: Double) = mix(this, other, factor)
-
-    override fun toVector4(): Vector4 = Vector4(x, s, l, alpha)
-}
-
-fun ColorRGBa.toXSLUVa() = toHSLUVa().toXSLUVa()
-
-fun mix(left: ColorXSLUVa, right: ColorXSLUVa, x: Double): ColorXSLUVa {
-    val sx = x.coerceIn(0.0, 1.0)
-    return ColorXSLUVa(
-            mixAngle(left.x, right.x, sx),
-            (1.0 - sx) * left.s + sx * right.s,
-            (1.0 - sx) * left.l + sx * right.l,
-            (1.0 - sx) * left.alpha + sx * right.alpha)
-}
-
-private fun map(x: Double, a: Double, b: Double, c: Double, d: Double): Double {
+internal fun map(x: Double, a: Double, b: Double, c: Double, d: Double): Double {
     return ((x - a) / (b - a)) * (d - c) + c
 }
 
@@ -211,90 +166,6 @@ private fun hueToX(hue:Double): Double {
     }
 }
 
-private fun xToHue(x:Double) : Double {
-    val x = x % 360.0
-    return if (0.0 <= x && x < 60.0) {
-        map(x, 0.0, 60.0, 0.0, 35.0)
-    } else if (60.0 <= x && x < 120.0) {
-        map(x, 60.0, 120.0, 35.0, 60.0)
-    } else if (120.0 <= x && x < 180.0) {
-        map(x, 120.0, 180.0, 60.0, 135.0)
-    } else if (180.0 <= x && x < 240.0) {
-        map(x, 180.0, 240.0, 135.0, 225.0)
-    } else if (240.0 <= x && x < 300.0) {
-        map(x, 240.0, 300.0, 225.0, 275.0)
-    } else {
-        map(x, 300.0, 360.0, 276.0, 360.0)
-    }
-}
-
-data class ColorHPLUVa(val h: Double, val s: Double, val l: Double, override val alpha: Double = 1.0) :
-        ColorModel<ColorHPLUVa>,
-        HueShiftableColor<ColorHPLUVa>,
-        SaturatableColor<ColorHPLUVa>,
-        ShadableColor<ColorHPLUVa>,
-        AlgebraicColor<ColorHPLUVa> {
-    fun toLCHUVa(): ColorLCHUVa {
-        val l1 = l
-        if (l1 > 0.9999999) {
-            return ColorLCHUVa(100.0, 0.0, h)
-        }
-        if (l1 < 0.00000001) {
-            return ColorLCHUVa(0.0, 0.0, h)
-        }
-        val l100 = l1 * 100.0
-        val max100 = maxSafeChromaForL(l100)
-        val c100 = max100 * s
-        return ColorLCHUVa(l100, c100, h)
-    }
-
-    override fun shiftHue(shiftInDegrees: Double): ColorHPLUVa {
-        return copy(h = h + (shiftInDegrees))
-    }
-
-    override fun shade(factor: Double): ColorHPLUVa = copy(l = l * factor)
-
-    override fun saturate(factor: Double): ColorHPLUVa = copy(s = s * factor)
-
-    override fun toRGBa(): ColorRGBa = toLCHUVa().toRGBa()
-
-    override fun opacify(factor: Double) = copy(alpha = alpha * factor)
-
-    override fun minus(right: ColorHPLUVa) = copy(h = h - right.h, s = s - right.s, l = l - right.l, alpha = alpha - right.alpha)
-
-    override fun plus(right: ColorHPLUVa) = copy(h = h + right.h, s = s + right.s, l = l + right.l, alpha = alpha + right.alpha)
-
-    override fun times(scale: Double) = copy(h = h * scale, s = s * scale, l = l * scale, alpha = alpha * scale)
-
-    override fun mix(other: ColorHPLUVa, factor: Double) = mix(this, other, factor)
-
-    override fun toVector4(): Vector4 = Vector4(h, s, l, alpha)
-}
-
-fun mix(left: ColorHPLUVa, right: ColorHPLUVa, x: Double): ColorHPLUVa {
-    val sx = x.coerceIn(0.0, 1.0)
-    return ColorHPLUVa(
-            mixAngle(left.h, right.h, sx),
-            (1.0 - sx) * left.s + sx * right.s,
-            (1.0 - sx) * left.l + sx * right.l,
-            (1.0 - sx) * left.alpha + sx * right.alpha)
-}
-
-
-fun ColorLCHUVa.toHPLUVa(): ColorHPLUVa {
-    val l100 = l
-    if (l100 > 99.9999999) {
-        return ColorHPLUVa(h, 0.0, 1.0)
-    }
-    if (l100 < 0.00000001) {
-        return ColorHPLUVa(h, 0.0, 0.0)
-
-    }
-    val max100 = maxSafeChromaForL(l)
-    val s1 = c / max100
-    return ColorHPLUVa(h, s1, l100 / 100.0)
-}
-
 fun ColorLCHUVa.toHSLUVa(): ColorHSLUVa {
     val l100 = l
 
@@ -311,4 +182,3 @@ fun ColorLCHUVa.toHSLUVa(): ColorHSLUVa {
 }
 
 fun ColorRGBa.toHSLUVa(): ColorHSLUVa = toLCHUVa().toHSLUVa()
-fun ColorRGBa.toHPLUVa(): ColorHPLUVa = toLCHUVa().toHPLUVa()
